feat(validations): 최소 주문 금액 유효성 검증 로직 구현 및 테스트
